Move parsePrivilege to MetastoreUtil
